back/pyproject.toml
[project]
name = "Gitlab_Reviewer_backend"
version = "1.0"
description = "基于 AI 的 GitLab 自动代码评审系统。使用人工智能对私有化gitlab中的仓库、提交、合并请求进行评审"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=1.93.0",
    "pydantic-settings>=2.10.1",
    "pymysql>=1.1.1",
    "python-gitlab>=6.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
